Add files via upload
The shooting method "shoots" for the c(0) initial condition by using the secant method to locate the c(0) the yields the known k(0). The value for k(0) is treated as a "zero" in the sense that we set a function f(k) = k-k0; note that f(k0) = 0. The secant method then uses cu = cl - fl*(cu-cl)/(fu-fl) to locate the next iterate for a c starting value that shoots for c(0). Note that the "l" and "u" stand for "lower" and "upper" respectively. The only "guess" required is cl, which can be chosen as cinf-eps where eps is some small perturbation such as 1e-4.