Infinite horizon reverse time shooting method
Made the shooting method for the full commitment case more efficient by removing two of the IVP solver statements and recycling the consumption and capital variables from previous iterations. The no commitment case is solved using the time-elimination method. The system is linearized about the steady-state and the eigenvalues and eigenvectors of the Jacobian matrix are computed to analyze the stability of the linearized system.